backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
pymongo
motor
clerk-backend-api
sendgrid
openai
pypdf
python-docx
httpx
# AI Services
anthropic  # Fallback AI provider

# Dependencies that were missing and caused startup failures
python-multipart
pydantic[email]
svix
lxml
cryptography